Update README.md to explain difference between 'on' and 'once' button event handlers
This one tripped me up a bit.  I was triggering the same ReactiveModal object in multiple callbacks in my application code, each with different 'on' button callbacks.  I was seeing multiple callbacks firing when I clicked the button on the reactive modal.  Finally realised that 'on' was adding additional handlers.  I switched to 'once', and everything worked as expected.